% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\name{ibkr_portfolio_accounts}
\alias{ibkr_portfolio_accounts}
\title{Get all accounts associated with the authenticated user}
\usage{
ibkr_portfolio_accounts()
}
\value{
Data frame with one row per account and columns: \code{account_id},
\code{type}, \code{currency}, \code{alias}
}
\description{
Get all accounts associated with the authenticated user
}
