% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.r
\name{ibkr_reauthenticate}
\alias{ibkr_reauthenticate}
\title{Reauthenticate the session}
\usage{
ibkr_reauthenticate()
}
\value{
Invisibly returns the response list
}
\description{
Call this if the session has timed out but the gateway is still running.
After calling this function, wait a few seconds then call \code{ibkr_tickle()}
to confirm the session is restored.
}
